requirements.txt
Flask
Flask-Cors
scikit-learn
pandas
numpy
mysql-connector-python
joblib